distsqlrun: use uint64 for offsets in window frames
Previously, we stored integer offsets of window frame boundaries
as uint32 which could lead to incorrect behavior in some very
unlucky edge cases (e.g. rank() OVER (ROWS MaxInt32+1 PRECEDING)
or similar would overflow and be treated as something like
ROWS 0 PRECEDING - the correct behavior would be to include up to
about 2 billions of preceding rows while we would include only the
current one). Now, we use uint64, so such cases will be handled
correctly.

Note that uint32 and uint64 are compatible, so there is no need
to bump the DistSQL version.

Release note: None

+++ pkg/sql/distsqlpb/processors.go
@@ -99,11 +99,11 @@ func (spec *WindowerSpec_Frame_Bounds) initFromAST(
 		}
 		switch m {
 		case tree.ROWS:
-			startOffset := int(tree.MustBeDInt(dStartOffset))
+			startOffset := int64(tree.MustBeDInt(dStartOffset))
 			if startOffset < 0 {
 				return pgerror.Newf(pgerror.CodeInvalidWindowFrameOffsetError, "frame starting offset must not be negative")
 			}
-			spec.Start.IntOffset = uint32(startOffset)
+			spec.Start.IntOffset = uint64(startOffset)
 		case tree.RANGE:
 			if isNegative(evalCtx, dStartOffset) {
 				return pgerror.Newf(pgerror.CodeInvalidWindowFrameOffsetError, "invalid preceding or following size in window function")
@@ -119,11 +119,11 @@ func (spec *WindowerSpec_Frame_Bounds) initFromAST(
 			}
 			spec.Start.TypedOffset = buf
 		case tree.GROUPS:
-			startOffset := int(tree.MustBeDInt(dStartOffset))
+			startOffset := int64(tree.MustBeDInt(dStartOffset))
 			if startOffset < 0 {
 				return pgerror.Newf(pgerror.CodeInvalidWindowFrameOffsetError, "frame starting offset must not be negative")
 			}
-			spec.Start.IntOffset = uint32(startOffset)
+			spec.Start.IntOffset = uint64(startOffset)
 		}
 	}
 
@@ -141,11 +141,11 @@ func (spec *WindowerSpec_Frame_Bounds) initFromAST(
 			}
 			switch m {
 			case tree.ROWS:
-				endOffset := int(tree.MustBeDInt(dEndOffset))
+				endOffset := int64(tree.MustBeDInt(dEndOffset))
 				if endOffset < 0 {
 					return pgerror.Newf(pgerror.CodeInvalidWindowFrameOffsetError, "frame ending offset must not be negative")
 				}
-				spec.End.IntOffset = uint32(endOffset)
+				spec.End.IntOffset = uint64(endOffset)
 			case tree.RANGE:
 				if isNegative(evalCtx, dEndOffset) {
 					return pgerror.Newf(pgerror.CodeInvalidWindowFrameOffsetError, "invalid preceding or following size in window function")
@@ -161,11 +161,11 @@ func (spec *WindowerSpec_Frame_Bounds) initFromAST(
 				}
 				spec.End.TypedOffset = buf
 			case tree.GROUPS:
-				endOffset := int(tree.MustBeDInt(dEndOffset))
+				endOffset := int64(tree.MustBeDInt(dEndOffset))
 				if endOffset < 0 {
 					return pgerror.Newf(pgerror.CodeInvalidWindowFrameOffsetError, "frame ending offset must not be negative")
 				}
-				spec.End.IntOffset = uint32(endOffset)
+				spec.End.IntOffset = uint64(endOffset)
 			}
 		}
 	}
